Modules documentation and rename `config.utils` submodule to `config.parser`
